triage-microservice/requirements.txt
Flask==3.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
